// src/lib/api/generated/models/postCategory.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Thisiscz API
 * Thisiscz 后端 API 文档
 * OpenAPI spec version: v1
 */

export type PostCategory = (typeof PostCategory)[keyof typeof PostCategory]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PostCategory = {
  life: 'life',
  work: 'work',
  crypto: 'crypto',
  sports: 'sports',
} as const